+++ graylog2-server/src/main/java/org/graylog2/bindings/ElasticsearchModule.java
@@ -16,10 +16,13 @@
  */
 package org.graylog2.bindings;
 
+import com.github.zafarkhaja.semver.Version;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.inject.AbstractModule;
 import io.searchbox.client.JestClient;
+import org.graylog2.bindings.annotations.ElasticsearchVersion;
+import org.graylog2.bindings.providers.ElasticsearchVersionProvider;
 import org.graylog2.bindings.providers.JestClientProvider;
 import org.graylog2.indexer.IndexMappingFactory;
 
@@ -29,5 +32,6 @@ public class ElasticsearchModule extends AbstractModule {
         bind(Gson.class).toInstance(new GsonBuilder().create());
         bind(JestClient.class).toProvider(JestClientProvider.class).asEagerSingleton();
         bind(IndexMappingFactory.class).asEagerSingleton();
+        bind(Version.class).annotatedWith(ElasticsearchVersion.class).toProvider(ElasticsearchVersionProvider.class);
     }
 }
