+++ graylog2-server/src/main/java/org/graylog2/bindings/annotations/ElasticsearchVersion.java
@@ -1,0 +1,4 @@
+package org.graylog2.bindings.annotations;
+
+public @interface ElasticsearchVersion {
+}
