+++ graylog2-server/src/main/java/org/graylog2/indexer/IndexMappingFactory.java
@@ -17,46 +17,21 @@
 package org.graylog2.indexer;
 
 import com.github.zafarkhaja.semver.Version;
-import io.searchbox.client.JestClient;
-import io.searchbox.client.JestResult;
-import io.searchbox.core.Ping;
-import org.graylog2.indexer.cluster.jest.JestUtils;
-import org.graylog2.indexer.gson.GsonUtils;
+import org.graylog2.bindings.annotations.ElasticsearchVersion;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
-import java.util.Optional;
-
-import static java.util.Objects.requireNonNull;
 
 @Singleton
 public class IndexMappingFactory {
-    private final JestClient jestClient;
+    private final Version elasticsearchVersion;
 
     @Inject
-    public IndexMappingFactory(JestClient jestClient) {
-        this.jestClient = requireNonNull(jestClient, "jestClient");
-    }
-
-    private Optional<Version> getElasticsearchVersion() {
-        final Ping request = new Ping.Builder().build();
-        final JestResult jestResult = JestUtils.execute(jestClient, request, () -> "Unable to retrieve Elasticsearch version");
-        return Optional.ofNullable(jestResult.getJsonObject())
-                .map(json -> GsonUtils.asJsonObject(json.get("version")))
-                .map(json -> GsonUtils.asString(json.get("number")))
-                .map(this::parseVersion);
-    }
-
-    private Version parseVersion(String version) {
-        try {
-            return Version.valueOf(version);
-        } catch (Exception e) {
-            throw new ElasticsearchException("Unable to parse Elasticsearch version: " + version, e);
-        }
+    public IndexMappingFactory(@ElasticsearchVersion Version elasticsearchVersion) {
+        this.elasticsearchVersion = elasticsearchVersion;
     }
 
     public IndexMapping createIndexMapping() {
-        final Version elasticsearchVersion = getElasticsearchVersion().orElseThrow(() -> new ElasticsearchException("Unable to retrieve Elasticsearch version"));
         if (elasticsearchVersion.satisfies(">=2.1.0 & <5.0.0")) {
             return new IndexMapping2();
         } else if (elasticsearchVersion.satisfies("^5.0.0")) {
